% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mutations.R
\name{add_depths_to_ref}
\alias{add_depths_to_ref}
\title{Add Read Depths To Ref}
\usage{
add_depths_to_ref(ref, samp, samp.dir)
}
\arguments{
\item{ref}{reference genome in "MixVir" format (genomic positions repeated for each associated feature they're associated with, etc.)}

\item{samp}{Character vector length 1 storing the name of a csv file with data for one sample. Should have columns CHROM,POS,REF,ALT,DP,AF. Others will be ignored.}

\item{samp.dir}{Directory storing sample file defined with parameter 'samp'.}
}
\value{
Data frame with cols "genomic_pos"	"ref_base"	"gene"	"ref_codon"	"ref_AA	gene_aa_position"	"ref_identity" "DP"
}
\description{
Add column for total read depths for a given sample to reference df
}
\examples{
add_depths_to_ref()
}
\keyword{depth}
