% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mutations.R
\name{bulk_vcf_to_mixvir}
\alias{bulk_vcf_to_mixvir}
\title{Convert VCF Samples to MixviR Input Format}
\usage{
bulk_vcf_to_mixvir(vcf.dir = NULL, csv.path = NULL, max.vcf.size = 1e+08)
}
\arguments{
\item{vcf.dir}{path to directory containing one or more vcf files that will be converted to csv input format for MixviR call_mutations(). VCF's need to contain "DP" and "AD" flags in the FORMAT field. This directory should not contain any other files.}

\item{csv.path}{Path to directory where csv files will be written.}

\item{max.vcf.size}{Max memory usage (in bytes) allowed when reading in vcf file (from vcfR).}
}
\value{
csv file(s) with cols "CHR"	"POS"	"REF"	"ALT"	"DP"	"REF_COUNT" ALT_COUNT". Written to path defined with 'csv.path'.
}
\description{
Create csv files with relevant contents of VCF to use as input for MixviR. Only necessary for early versions of MixviR.
}
\keyword{VCF}
