% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mutations.R
\name{call_mutations}
\alias{call_mutations}
\title{Identify Sample Amino Acid Changes}
\usage{
call_mutations(
  sample.dir,
  min.alt.freq = 0.01,
  name.sep = "NULL",
  reference = "https://raw.githubusercontent.com/mikesovic/IDI-AMSL/main/SC2_ref.tsv",
  write.mut.table = FALSE
)
}
\arguments{
\item{sample.dir}{Path to directory with one tab-delimited file for each sample to analyze.
Each file should contain columns named POS, REF, ALT, AF, DP and is generally
a summary from a vcf file. Additional columns can be included and will be ignored.
POS = genomic position, REF = reference base, ALT = alternate base/allele, AF = alt frequency, DP = read depth at site}

\item{min.alt.freq}{Minimum frequency (0-1) for retaining alternate allele. Default = 0.01.}

\item{name.sep}{Character in sample names that separates the unique sample identifier (characters preceeding the separator) from any additional text. Only text preceeding the first instance of the character will be retained.}

\item{reference}{Reference genome information in MixVir format.}

\item{write.mut.table}{Logical to indicated whether to write a text file that stores all mutations
for all samples analyzed to working directory. This is the same information contained in data frame
'samp_mutations' created by this function. Default = FALSE}
}
\value{
Data frame 'samp_mutations' containing amino acid changes observed for each sample.
}
\description{
Function to identify full set of amino acid changes in a given sample (includes changes based on both SNVs and indels)
}
\examples{
call_mutations()
}
\keyword{mutation}
