% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ref.R
\name{create_ref}
\alias{create_ref}
\title{Create MixVir reference genome object}
\usage{
create_ref(genome, feature.bed)
}
\arguments{
\item{genome}{fasta formatted genome file}

\item{feature.bed}{bed file defining features of interest (open reading frames to translate). Tab delimited with 4 columns (without column names):"chr", "start", "end", "feature_name".}
}
\value{
A data frame with columns CHR,POS,REF_BASE,GENE,REF_CODON,REF_AA,GENE_AA_POS,REF_IDENT
}
\description{
Uses a fasta genome and bed file defining features of interest (genes/ORFs) to create a data frame that's used as a reference to translate nucleotide data to amino acids and subsequently call variants/mutations from a sample.
}
\examples{
create_ref()
}
\keyword{reference}
