% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{explore_mutations}
\alias{explore_mutations}
\title{MixviR Shiny Dashboard}
\usage{
explore_mutations(dates = NULL, lineage.muts = NULL)
}
\arguments{
\item{dates}{path to csv file with cols "SAMP_NAME" and "date". Sample names need to match those in samp_mutations data frame created by call_mutations().}

\item{lineage.muts}{path to csv file with cols "Gene", "ALT_ID", "Lineage" containing mutations associated with lineages of interest. See example file at "https://github.com/mikesovic/MixviR/blob/main/mutation_files/outbreak_20211202.csv". Can use this example file by setting lineage.muts to "https://raw.githubusercontent.com/mikesovic/MixviR/main/mutation_files/outbreak_20211202.csv"}
}
\value{
Shiny Dashboard to Explore Data
}
\description{
Open Dashboard To Explore Mutation Data
}
\examples{
explore_mutations()
}
\keyword{shiny}
