% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mutations.R
\name{id_indels}
\alias{id_indels}
\title{ID Indel-based Amino Acid Changes}
\usage{
id_indels(variant.calls, ref)
}
\arguments{
\item{variant.calls}{Data frame with cols POS, REF, ALT, AF, DP. Additional columns will be ignored.}

\item{ref}{reference genome in "MixVir" format (from create_ref() function)}
}
\value{
Data frame that includes amino acid calls based on indel variants observed in sample. Contains cols "POS", "REF_BASE", "GENE", "REF_CODON", "REF_AA", "GENE_AA_POS", "REF_IDENT", "REF", "ALT", "ALT_freq", "ALT_COUNT", "samp_codon", "samp_AA", "samp_identity", "DP"
}
\description{
Identify amino acid changes associated with indel variation. Changes associated with SNVs are identified in separate function. Used by call_mutations() function.
}
\keyword{indel}
