% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prop_char_muts.R
\name{plot_prop_charac_muts}
\alias{plot_prop_charac_muts}
\title{Plot Freqencies Of Lineage-Characteristic Mutations}
\usage{
plot_prop_charac_muts(
  genes = "all",
  min.read.depth = 50,
  lineage.def = "CDC",
  variant.thresh = 0.7,
  grid.dim = c(1, 1),
  lineage.def.name = NULL
)
}
\arguments{
\item{genes}{Character vector defining genes to include in analysis. Default = "all".}

\item{min.read.depth}{Minimum number of reads required at the site to include (currently total reads)}

\item{lineage.def}{Tab-delimited file containing an a priori set of amino acid changes associated with variants/lineages of interest. Amino acid changes in the sample are matched to these. Choose from 'CDC' (default), 'Outbreak', 'Outbreak_May2021', or define your own.}

\item{variant.thresh}{Value (0-1) that defines horizontal reference line on plots}

\item{grid.dim}{Defines number of rows and columns plots for individual samples will be divided into for a given variant/lineage. Default: "c(1,1)" gives one plot per page.}

\item{lineage.def.name}{Name to use in plot titles to define the source of the lineage/variant definitions. Defaults to name for lineage.def argument.}
}
\description{
For each variant/lineage+sample combination, generates a barplot representing frequencies of each amino acid change associated with the lineage/variant.
}
\examples{
plot_prop_charac_muts()
}
\keyword{plot}
