% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odh_table.R
\name{write_ODH_table}
\alias{write_ODH_table}
\title{Write Table In New ODH Format}
\usage{
write_ODH_table(
  sample.dir,
 
    mut.positions = "https://raw.githubusercontent.com/mikesovic/IDI-AMSL/main/ODH_mut_pos_master_split_dels_6-3-21.txt",
  name.sep = "NULL",
  reference = "https://raw.githubusercontent.com/mikesovic/IDI-AMSL/main/SC2_ref.tsv",
 
    site.meta = "https://raw.githubusercontent.com/mikesovic/IDI-AMSL/main/OSU_SequencingSites_CODEX.csv"
)
}
\arguments{
\item{sample.dir}{Path to directory with one tab-delimited file for each sample to analyze.
Each file should contain columns named POS, REF, ALT, AF, DP and is generally
a summary from a vcf file. Additional columns can be included and will be ignored.
POS = genomic position, REF = reference base, ALT = alternate base/allele, AF = alt frequency, DP = read depth at site}

\item{mut.positions}{Data frame that defines amino acid changes of interest and associated mutation positions}

\item{name.sep}{Character in sample names that separates the unique sample identifier (characters preceeding the separator) from any additional text. Only text preceeding the first instance of the character will be retained.}

\item{reference}{Reference genome information in MixVir format.}

\item{site.meta}{Comma separated table with metadata for each sample site. Cols codes, Site_ID, Facility_Name, LAB, replicate
Should have two columns - 'MUT' and 'POS'}
}
\description{
Write a table with read counts for each position
}
\examples{
write_ODH_table()
}
\keyword{ODH}
