% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mutations.R
\name{add_depths_to_ref}
\alias{add_depths_to_ref}
\title{Add Read Depths To Ref}
\usage{
add_depths_to_ref(ref, samp.variants)
}
\arguments{
\item{ref}{reference genome in "MixVir" format (from create_ref() function)}

\item{samp.variants}{Data frame produced by function vcf_to_mixvir(). Contains columns "CHR", "POS", "REF", "ALT", "DP", "ALT_COUNT".}
}
\value{
Original 'ref' data frame with depth (DP) column added: cols "genomic_pos"	"ref_base"	"gene"	"ref_codon"	"ref_AA"	"GENE_AA_POS"	"ref_identity" "DP"
}
\description{
Add column of total read depths for a given sample to reference df. Run as part of id_mutations().
}
\keyword{depth}
