% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ref.R
\name{create_ref}
\alias{create_ref}
\title{Create MixVir reference genome object}
\usage{
create_ref(genome, feature.bed)
}
\arguments{
\item{genome}{fasta formatted genome file (assumes only one chromosome)}

\item{feature.bed}{bed file (4 columns, tab delimited, no column names: chr, start, end, feature_name) defining features of interest (open reading frames to translate)}
}
\description{
Uses a fasta genome (single chromosome) and bed file defining features of interest (genes) to create a data frame used as a reference to call variants/mutations from a sample.
}
\examples{
create_ref()
}
\keyword{reference}
