% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{explore_mutations}
\alias{explore_mutations}
\title{MixviR Shiny Dashboard}
\usage{
explore_mutations(
  dates = NULL,
  lineage.muts = NULL,
  read.muts.from = NULL,
  all.target.muts = FALSE
)
}
\arguments{
\item{dates}{Path to optional csv file with cols "SAMP_NAME", "LOCATION", and "DATE". Sample names need to match those in \emph{samp_mutations} data frame created by \code{call_mutations()}. Dates should be provided in the format \emph{mmddyyyy}.}

\item{lineage.muts}{Path to optional csv file with required cols "Gene", "Mutation", and  "Lineage" defining mutations associated with lineages of interest. See example file at "https://github.com/mikesovic/MixviR/blob/main/mutation_files/outbreak_20211202.csv". Additional columns will be ignored.}

\item{read.muts.from}{By default, data are read from \emph{samp_mutations} data frame that is created by \code{call_mutations()} and written to the global environment. If this data frame has been written to a file (see \emph{write.mut.table} in \code{call_mutations()}), the dashboard can optionally be opened by providing the path to that file. The fields "SAMP_NAME, CHR, POS, GENE, ALT_ID, AF, DP" must be present (additional fields will be ignored).}

\item{all.target.muts}{Logical to indicate whether results for all target mutations were written to the output of the \code{call_mutations()} function. See \emph{write.all.targets} option in \code{call_mutations()}. Default FALSE. If TRUE, more informative sequencing depth information can be provided in dashboard.}
}
\value{
Shiny Dashboard to Explore Data
}
\description{
Open dashboard to explore mutation data generated with \code{call_mutations()}.
}
\examples{
if (interactive()) {explore_mutations(read.muts.from = system.file("extdata", 
                                                                   "sample_mutations.tsv", 
                                                                   package = "MixviR"),
                                      lineage.muts = system.file("extdata", 
                                                                 "example_lineage_muts.csv", 
                                                                 package = "MixviR"))}
if (interactive()) {explore_mutations(read.muts.from = system.file("extdata", 
                                                                   "sample_mutations.tsv", 
                                                                   package = "MixviR"), 
                                      dates = system.file("extdata", 
                                                          "example_location_date.csv", 
                                                          package = "MixviR"), 
                                      lineage.muts = system.file("extdata", 
                                                                 "example_lineage_muts.csv", 
                                                                 package = "MixviR"))}
}
\keyword{shiny}
