% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mutations.R
\name{vcf_to_mixvir}
\alias{vcf_to_mixvir}
\title{Convert Sample VCF to MixviR Input Format}
\usage{
vcf_to_mixvir(infile, max.vcf.size = 1e+08)
}
\arguments{
\item{infile}{path to a vcf file}

\item{max.vcf.size}{Max memory usage (in bytes) allowed when reading in vcf file (from vcfR)}
}
\value{
Data frame with cols "CHR"	"POS"	"REF"	"ALT"	"DP" "REF_COUNT"	"ALT_COUNT"
}
\description{
Create data frame with relevant contents of VCF
}
\examples{
vcf_to_mixvir()
}
\keyword{VCF}
